\name{as.afactor}
\alias{as.afactor}
\title{As a factor}
\description{
 Changes columns of a data frame into factors
}
\usage{
	as.afactor(cName,data)
}

\arguments{
	\item{cName}{The quoted column name to change}
	\item{data}{A data frame, or matrix. }
}
\value{
	A data frame with changes in column cName. 
}


\author{
	Bob Wheeler \email{bwheelerg@gmail.com}
	
	Please cite this program as follows:

Wheeler, R.E. (2009). as.twolevel \emph{RelativeRisk}. The R project for statistical computing \url{http://www.r-project.org/}
}

\keyword{htest}
