\name{Chocolates}
\alias{Chocolates}
\title{Chocolates prferences}
\description{
 A Data Set illustrating a simple 0 or 1 response.
}
\usage{
	data(Chocolates)
}

\details{

	A frequently cited data set for consumer preferences.  

}

\author{
	Bob Wheeler \email{bwheelerg@gmail.com}
	}

\keyword{datasets}

