\name{Melanoma}
\alias{Melanoma}
\title{Primary cutaneous malignant melanoma.}
\description{
 A Data Set illustrating large odds ratios
}
\usage{
	data(Melanoma)
}

\details{

	Contingency table data from Veronesi et.al. (1988).  

}
\references{
  Veronesi U., Cascinelli N.,Adamus J., et.al. Thin stage I primary
  cutaneous malignant melanoma. \emph{N Engl. J. Med.}. 1988. 318. 1159-62.
  }

\author{
	Bob Wheeler \email{bwheelerg@gmail.com}
	}

\keyword{datasets}

