% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bactdate.R
\name{plt_gubbin_tree}
\alias{plt_gubbin_tree}
\title{Plot Gubbin Tree}
\usage{
plt_gubbin_tree(gubbin_tree)
}
\arguments{
\item{gubbin_tree}{tree file generated by Gubbins}

\item{bactdate_tree}{tree file generated from bactdating}

\item{coltime_range}{range of collection time}

\item{divtime_range}{range of divergence time}

\item{distance_range}{range of the distance}
}
\value{
plot of Gubbin tree
}
\description{
This function generates Gubbin Tree plot.
}
