% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RelativeRiskFunction.R
\name{PltRR}
\alias{PltRR}
\title{Generate Pairwise Matrix for Relative Risk Calculation}
\usage{
PltRR(rrData, RRplotTitle)
}
\arguments{
\item{rrData}{relative risk generated by RelativeRisk()}

\item{RRplotTitle}{name of the plot}
}
\value{
relative risk plot
}
\description{
This function generates pairwise matrices from metadata file.
}
