% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RelativeRiskFunction.R
\name{ratio_bootstrap}
\alias{ratio_bootstrap}
\title{Generate Pairwise Matrix for Relative Risk Calculation}
\usage{
ratio_bootstrap(x, geo_matrix, colyear_matrix, strain_matrix)
}
\arguments{
\item{x}{selected index}

\item{geo_matrix}{pairwise geographical matrix generated by RR_matrix() function}

\item{colyear_matrix}{pairwise collection year matrix generated by RR_matrix() function}

\item{strain_matrix}{pairwise lineage matrix generated by RR_matrix() function}
}
\value{
relative risk
}
\description{
This function generates pairwise matrices from metadata file.
}
