% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RelativeRiskFunction.R
\name{RelativeRisk}
\alias{RelativeRisk}
\title{Generate Pairwise Matrix for Relative Risk Calculation}
\usage{
RelativeRisk(geo_matrix, colyear_matrix, strain_matrix, metadataset)
}
\arguments{
\item{geo_matrix}{pairwise geographical matrix generated by RR_matrix() function}

\item{colyear_matrix}{pairwise collection year matrix generated by RR_matrix() function}

\item{strain_matrix}{pairwise lineage matrix generated by RR_matrix() function}

\item{metadataset}{metadata}
}
\value{
relative risk with bootstrap
}
\description{
This function generates pairwise matrices from metadata file.
}
