% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RelativeRiskFunction.R
\name{RR_matrix}
\alias{RR_matrix}
\title{Generate Pairwise Matrix for Relative Risk Calculation}
\usage{
RR_matrix(
  meta_table,
  bactdate_tree,
  coltime_range = 1,
  divtime_range = 50,
  distance_range = 100
)
}
\arguments{
\item{meta_table}{metadata file}

\item{bactdate_tree}{tree file generated from bactdating}

\item{coltime_range}{range of collection time}

\item{divtime_range}{range of divergence time}

\item{distance_range}{range of the distance}
}
\value{
pairwise matrices
}
\description{
This function generates pairwise matrices from metadata file.
}
