\name{simData}
\alias{simData}
\title{Simulated Case-Control Data}
\description{
 A Simulation Data Set
}
\usage{
	data(simData)
}

\details{

	Simulated data for 600 cases and 200 controls from a population with a marginal case probability of
	\eqn{P=0.3}, and true relative risk values (0.3, 1.5, 2.0, 0.7, 0.3). 

}


\author{
	Bob Wheeler \email{bwheelerg@gmail.com}
	}

\keyword{datasets}

