% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bactdate.R
\name{time_resolved}
\alias{time_resolved}
\title{Divergence time estimated by Bactdating}
\usage{
time_resolved(
  match_tree_meta,
  dropRef_tree,
  save_bd_path,
  time = "Year_collection"
)
}
\arguments{
\item{match_tree_meta}{metadata file of combining tree file and metafile}

\item{dropRef_tree}{tree file generated from Gubbins without reference tip}

\item{save_bd_path}{path to save bactdating results}

\item{time}{use "Year_Collection" by default}
}
\value{
roottotip and bactdating results
}
\description{
This function generates estimated divergence time between isolates
}
